+++ audio-ui/src/androidTest/java/com/google/android/horologist/audio/ui/VolumeScreenTest.kt
@@ -17,22 +17,16 @@
 package com.google.android.horologist.audio.ui
 
 import android.os.Vibrator
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.focus.FocusRequester
 import androidx.compose.ui.focus.focusRequester
-import androidx.compose.ui.platform.LocalConfiguration
-import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.test.ExperimentalTestApi
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onRoot
 import androidx.compose.ui.test.performRotaryScrollInput
 import androidx.test.filters.MediumTest
 import androidx.test.platform.app.InstrumentationRegistry
-import androidx.wear.compose.material.Text
 import com.google.android.horologist.audio.ExperimentalHorologistAudioApi
 import com.google.android.horologist.audio.VolumeState
 import com.google.common.truth.Truth.assertThat
@@ -40,7 +34,12 @@ import kotlinx.coroutines.test.runTest
 import org.junit.Rule
 import org.junit.Test
 
-@OptIn(ExperimentalHorologistAudioApi::class, ExperimentalHorologistAudioUiApi::class, ExperimentalTestApi::class, kotlinx.coroutines.ExperimentalCoroutinesApi::class)
+@OptIn(
+    ExperimentalHorologistAudioApi::class,
+    ExperimentalHorologistAudioUiApi::class,
+    ExperimentalTestApi::class,
+    kotlinx.coroutines.ExperimentalCoroutinesApi::class
+)
 @MediumTest
 class VolumeScreenTest {
     @get:Rule
@@ -48,7 +47,6 @@ class VolumeScreenTest {
 
     @Test
     fun testHaptics() = runTest {
-
         val rotaryPixelsForVolume = 136
         val volumeRepository = FakeVolumeRepository(VolumeState(50, 100))
         val audioOutputRepository = FakeAudioOutputRepository()
@@ -64,14 +62,12 @@ class VolumeScreenTest {
         val focusRequester = FocusRequester()
 
         composeTestRule.setContent {
-            val density = LocalDensity.current
-            val config = LocalConfiguration.current
-
-            Box(modifier = Modifier.fillMaxSize(), contentAlignment = Alignment.Center) {
-                Column {
-                    Text("densityDpi" + config.densityDpi)
-                }
-            }
+            VolumeScreen(
+                modifier = Modifier
+                    .fillMaxSize()
+                    .focusRequester(focusRequester),
+                model
+            )
         }
 
         composeTestRule.runOnIdle {
