+++ media-ui/src/main/java/com/google/android/horologist/media/ui/components/MediaControlButtons.kt
@@ -16,16 +16,9 @@
 
 package com.google.android.horologist.media.ui.components
 
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
 import androidx.wear.compose.material.ButtonColors
 import androidx.wear.compose.material.ButtonDefaults
 import com.google.android.horologist.media.ui.ExperimentalHorologistMediaUiApi
@@ -158,28 +151,4 @@ internal fun MediaControlButtons(
             )
         }
     )
-}
-
-@Composable
-public fun ControlButtonLayout(
-    leftButton: @Composable () -> Unit,
-    middleButton: @Composable () -> Unit,
-    rightButton: @Composable () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    Row(
-        modifier = modifier.fillMaxWidth(),
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.SpaceBetween
-    ) {
-        Box(modifier = Modifier.padding(start = 5.dp)) {
-            leftButton()
-        }
-
-        middleButton()
-
-        Box(modifier = Modifier.padding(end = 5.dp)) {
-            rightButton()
-        }
-    }
 }
